+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+- 5.2.1
+  - Fix type applied to filter funcs as another func's arg value (PR #743) (issue #586)
+  - Allow self nesting of functions with `allowSelfNesting` (PR #698)
 - 5.2.0
   - ! Breaking change: `children1` is now array in result of `getTree()` to preserve items order (PR #672) (issues #589, #670)
     `Utils.getTree(tree, true, false)` will behave same as before this change.